Fixed the way default color and font are used

+++ include/Color.h
@@ -6,10 +6,10 @@
 
 struct Color
 {
-    BYTE r;
-    BYTE g;
-    BYTE b;
-    BYTE a;
+    int r;
+    int g;
+    int b;
+    int a;
 
     Color()
     : r(0), g(0), b(0), a(0)
